JPA repository, dependency injection, database seeding

+++ src/main/java/com/narcsoft/course/entities/User.java
@@ -13,7 +13,7 @@ public class User implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
+	private Integer id;
 	private String name;
 	private String email;
 	private String phone;
@@ -22,7 +22,7 @@ public class User implements Serializable {
     public User() {
 }
 
-	public User(long id, String name, String email, String phone, String password) {
+	public User(Integer id, String name, String email, String phone, String password) {
 		super();
 		this.id = id;
 		this.name = name;
@@ -31,11 +31,11 @@ public class User implements Serializable {
 		this.password = password;
 	}
 
-	public long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
